New page factory, modules changes
