+++ bb-aem-64/src/main/java/com/cognifide/qa/bb/aem/core/siteadmin/SiteToolbar.java
@@ -17,17 +17,14 @@
  * limitations under the License.
  * #L%
  */
-package com.cognifide.qa.bb.aem.core.siteadmin.aem64;
+package com.cognifide.qa.bb.aem.core.siteadmin;
 
-import com.cognifide.qa.bb.constants.Timeouts;
-import com.cognifide.qa.bb.provider.selenium.BobcatWait;
 import com.cognifide.qa.bb.qualifier.PageObject;
 import com.google.inject.Inject;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 
 @PageObject(css = ".foundation-collection-actionbar")
 public class SiteToolbar {
@@ -35,13 +32,10 @@ public class SiteToolbar {
   @Inject
   private WebDriver webDriver;
 
-  @Inject
-  private BobcatWait bobcatWait;
-
   @FindBy(css = "button.granite-collection-create.foundation-toggleable-control.coral3-Button.coral3-Button--primary>coral-button-label")
   private WebElement createButton;
 
-  public void create(){
+  public void create() {
     Actions actions = new Actions(webDriver);
     actions.pause(2000).perform();
     createButton.click();
